Shuffle some stuff around into separate headers. I'll probably also write some more documentation.
