+++ docs/headers/celcd.rst
@@ -1,0 +1,19 @@
+.. _celcd_h:
+
+celcd.h
+======
+
+.. code-block:: c
+
+    #include <celcd.h>
+
+This header includes defines for the PL111 LCD controller used in the CE.
+
+.. contents:: :local:
+   :depth: 3
+
+API Documentation
+-----------------
+
+.. doxygenfile:: celcd.h
+   :project: CE C/C++ Toolchain
