+++ docs/headers/certc.rst
@@ -1,0 +1,19 @@
+.. _certc.h:
+
+certc.h
+======
+
+.. code-block:: c
+
+    #include <certc.h>
+
+This header includes defines for the Real-Time Clock (RTC) used in the CE.
+
+.. contents:: :local:
+   :depth: 3
+
+API Documentation
+-----------------
+
+.. doxygenfile:: certc.h
+   :project: CE C/C++ Toolchain
