+++ docs/headers/ceutil.rst
@@ -1,0 +1,19 @@
+.. _ceutil_h:
+
+ceutil.h
+======
+
+.. code-block:: c
+
+    #include <ceutil.h>
+
+This header includes defines for miscellaneous CE-specific things.
+
+.. contents:: :local:
+   :depth: 3
+
+API Documentation
+-----------------
+
+.. doxygenfile:: ceutil.h
+   :project: CE C/C++ Toolchain
