+++ docs/headers/tios.rst
@@ -1,0 +1,19 @@
+.. _tios_h:
+
+tios.h
+======
+
+.. code-block:: c
+
+    #include <tios.h>
+
+This header includes defines for OS memory and prototypes for OS routines.
+
+.. contents:: :local:
+   :depth: 3
+
+API Documentation
+-----------------
+
+.. doxygenfile:: tios.h
+   :project: CE C/C++ Toolchain
